rgw_sal_motr: milestone for RGW HSM support
IOSEA HSM utiity requires a S3 object is created as a Motr composite
object so that it can move the object's extents easily between tiers
(pools). To support HSM features, a few changes have been made:

(1) A configuration flag for HSM is added. HSM will be enabled if
    the flag is set (default is disabled).

(2) For the S3 objects which are created via PUT OBJ request, a composite
    object is created for it instead of a 'normal' Motr object.
    Now creation/deletion are working.

(3) For the S3 objects that are created and written via multipart upload,
    each S3 object is created as a composite object and each part is
    covered by an extent.

    currently, multipart upload (upload/abort/resume/deletion) with HSM
    support is also working.

    A known issue for multipart upload: only default part size (15MB) is
    supported, solution to support other part sizes will be done in later
    commits.